Change bg of list, make more mobile responsive
Still a lot of work to go for mobile responsiveness! Changed the background color on both the list of jokes and the side panel to be a transparent version of the main purple color. I also added a blur effect to the items underneat the transparent parts, for a frostrated glass effect.